Implement metadata util to centralize MetadataService client creation and provide a helper to get the instance's region.
