+++ lib/providers/token.js
@@ -2,9 +2,9 @@
 
 const Vaulted = require('vaulted');
 const promisify = require('./../utils/promisify');
-const AWS = require('aws-sdk');
 const http = require('http');
 const STATUS_CODES = require('./../control/util/status-codes');
+const metadata = require('./../utils/metadata').metadata;
 const preconditions = require('conditional');
 const checkNotNull = preconditions.checkNotNull;
 
@@ -18,8 +18,6 @@ const METADATA_ENDPOINTS = [
 const DEFAULT_VAULT_PORT = Config.get('vault:port');
 const DEFAULT_VAULT_HOST = Config.get('vault:host');
 const DEFAULT_VAULT_TLS = Config.get('vault:tls');
-
-const DEFAULT_METADATA_HOST = Config.get('metadata:host');
 
 // Default Warden connection options
 const DEFAULT_WARDEN_HOST = Config.get('warden:host');
@@ -58,23 +56,9 @@ class TokenProvider {
     const opts = options || {};
 
     // Set metadata options
-    let metadata = null,
-      client = null;
+    this._metadata = metadata(opts);
 
-    if (opts.metadata) {
-      opts.metadata.host = opts.metadata.host || DEFAULT_METADATA_HOST;
-
-      metadata = opts.metadata.client;
-    } else {
-      opts.metadata = {
-        host: DEFAULT_METADATA_HOST
-      };
-    }
-    if (!!metadata && metadata instanceof AWS.MetadataService) {
-      this._metadata = metadata;
-    } else {
-      this._metadata = new AWS.MetadataService({host: opts.metadata.host});
-    }
+    let client = null;
 
     // Set Vault connection options
     const vault = {
